use CLI flags for excluding repos
